+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [791](https://github.com/vegaprotocol/data-node/issues/791) - `v2` orders `api`
 
 ### 🐛 Fixes
+- [799](https://github.com/vegaprotocol/data-node/issues/799) - Latest versions of liquidity provision Should always be returned no matter the pagination order 
 - [](https://github.com/vegaprotocol/data-node/issues/xxx) -
 
 ## 0.53.0